requirements.txt
pytrends==4.9.0
openai==1.3.5
feedgen==0.9.0
python-dotenv==1.0.0
schedule==1.2.1
deep-translator==1.11.4
